Andando reinicio de contraseña - queda testear las otras funcionalidades

+++ Clase 19/controllers/products.controller.js
@@ -1,6 +1,7 @@
 import { productService } from "../repositories/services.js"
 import {faker} from '@faker-js/faker'
 import { logger } from "../logger/logger.js"
+import { CustomError } from "../errors/CustomError.js"
 
 const generateProducts = () => {
     return {
@@ -23,11 +24,11 @@ export class ProductController{
         const config = {limit:limit,pageQuery:pageQuery,sort:sort}
         const result = productService.getProducts(config)
         if (sort=="default"){
-            result.prevLink=result.hasPrevPage? req.protocol + '://' + req.get('host')+ "/?limit="+limit+"&page="+result.prevPage : result.hasPrevPage
-            result.nextLink=result.hasNextPage? req.protocol + '://' + req.get('host') +"/?limit="+limit+"&page="+result.nextPage : result.hasNextPage}
+            result.prevLink=result.hasPrevPage? req.protocol + '://' + req.get('host')+ "/Products?limit="+limit+"&page="+result.prevPage : result.hasPrevPage
+            result.nextLink=result.hasNextPage? req.protocol + '://' + req.get('host') +"/Products?limit="+limit+"&page="+result.nextPage : result.hasNextPage}
         else {
-                result.prevLink=result.hasPrevPage? req.protocol + '://' + req.get('host')+ "/?limit="+limit+"&page="+result.prevPage+"&sort="+sort : result.hasPrevPage
-                result.nextLink=result.hasNextPage? req.protocol + '://' + req.get('host') +"/?limit="+limit+"&page="+result.nextPage+"&sort="+sort : result.hasNextPage}
+                result.prevLink=result.hasPrevPage? req.protocol + '://' + req.get('host')+ "/Productslimit="+limit+"&page="+result.prevPage+"&sort="+sort : result.hasPrevPage
+                result.nextLink=result.hasNextPage? req.protocol + '://' + req.get('host') +"/Products?limit="+limit+"&page="+result.nextPage+"&sort="+sort : result.hasNextPage}
         res.status(200).json(result)
         }
         catch (err){
@@ -39,7 +40,12 @@ export class ProductController{
     try{
     const pid = req.params.pid
     const producto=productService.getProductById(pid)
-    producto? res.json(producto) : res.status(404).send('<h1>Error 404: El producto no existe</h1>')
+    producto? res.json(producto) : CustomError.createError({
+        name:"Product not found",
+        code:3,
+        cause:"The product does not exist in the current database",
+        message:`El producto ${id} no existe o no se encuentra en la base de datos actual`
+    })
     }
     catch(err){
         logger.error(err)
@@ -86,8 +92,18 @@ export class ProductController{
                     message:`Dato faltante o de tipo incorrecto\n Se recibió id=${typeof(id)},updateProduct=${typeof(updatedProduct)}}`
                 })
             }
-            result = productService.updateProduct(id=id,update=updatedProduct)
-            res.status(200).send(result)
+            const product = await productService.getProductById(pid=id)
+            if (product.owner == req.user.email || req.user.role=="admin"){
+                const result = await productService.updateProduct(id=id,update=updatedProduct)
+                res.status(200).send(result)
+            }
+            CustomError.createError({
+                name:"Invalid or missing params",
+                cause:"Needed params were wrong",
+                code:"2",
+                message:`Dato incorrecto\n Se recibió user=${req.user.email},no puedes actualizar un producto que no es tuyo}`
+            })
+            
         }
         catch(err){
             logger.error(err)
@@ -97,8 +113,17 @@ export class ProductController{
     deleteProduct= async(req, res) => {
         try{
             const id = req.params.pid
-            const result = productService.deleteProduct(id)
-            res.json(result)
+            const product = await productService.getProductById(pid=id)
+            if (product.owner == req.user.email || req.user.role=="admin"){
+                const result = await productService.deleteProduct(id)
+                res.status(200).send(result)
+            }
+            CustomError.createError({
+                name:"Invalid or missing params",
+                cause:"Needed params were wrong",
+                code:"2",
+                message:`Dato incorrecto\n Se recibió user=${req.user.email},no puedes borrar un product que no es tuyo}`
+            })
         }
             catch(err){
                     logger.error(err)
